Added a conditional with a warning for projects currently lacking a references (STANDARDS) folder

+++ src/smithlib/package.py
@@ -993,9 +993,12 @@ class differContext(Context.Context) :
         outputpath = getattr(Context.g_module, 'out', 'results')
         refpath = getattr(Context.g_module, 'STANDARDS', 'references')
         testspath = getattr(Context.g_module, 'TESTDIR', 'tests')
-        # need to find a way to get new wordlists from tests/ as param 
-        Utils.subprocess.Popen("diffenator2 diff --fonts-before " + outputpath + "/*.ttf " + "--fonts-after " + refpath + "/*.ttf " + "--out " + outputpath + "/diffenator2/", shell = 1).wait()
-        Utils.subprocess.Popen("rm -fv build.ninja .ninja_log", shell = 1).wait()
+        if not os.path.exists(refpath):
+            print("Stopping, no reference font files to diff against. Please add the font files from your last release to the {} folder.".format(refpath))
+        else:
+            # need to find a way to get new wordlists from tests/ as param using testpath 
+            Utils.subprocess.Popen("diffenator2 diff --fonts-before " + outputpath + "/*.ttf " + "--fonts-after " + refpath + "/*.ttf " + "--out " + outputpath + "/diffenator2/", shell = 1).wait()
+            Utils.subprocess.Popen("rm -fv build.ninja .ninja_log", shell = 1).wait()
 
 class graideContext(Build.BuildContext) :
     """Create graide .cfg files, one per font in graide/"""
